+++ src/manual/core.rs
@@ -1,3 +1,5 @@
+pub use CV_MAKETYPE as CV_MAKE_TYPE;
+
 pub use self::input_output_array::*;
 pub use self::mat::*;
 pub use self::point::*;
@@ -31,13 +33,13 @@ mod size;
 mod vec;
 
 #[inline(always)]
-pub const fn MAT_DEPTH(flags: i32) -> i32 {
+pub const fn CV_MAT_DEPTH(flags: i32) -> i32 {
     #![allow(non_snake_case)]
     flags & crate::core::Mat_DEPTH_MASK
 }
 
 #[inline(always)]
-pub const fn MAKETYPE(depth: i32, cn: i32) -> i32 {
+pub const fn CV_MAKETYPE(depth: i32, cn: i32) -> i32 {
     #![allow(non_snake_case)]
-    MAT_DEPTH(depth) + ((cn - 1) << crate::core::CV_CN_SHIFT)
+    CV_MAT_DEPTH(depth) + ((cn - 1) << crate::core::CV_CN_SHIFT)
 }