+++ tests/core.rs
@@ -1,5 +1,5 @@
 use opencv::{
-    core::{self, CV_32S, CV_32SC3, CV_64F, CV_64FC1, CV_8U, CV_8UC2, MAKETYPE, Moments, Point2f, RotatedRect, Scalar, Size2f},
+    core::{self, CV_32S, CV_32SC3, CV_64F, CV_64FC1, CV_8U, CV_8UC2, CV_MAKETYPE, Moments, Point2f, RotatedRect, Scalar, Size2f},
     prelude::*,
     Result,
     types::VectorOfMat,
@@ -7,9 +7,9 @@ use opencv::{
 
 #[test]
 fn make_type() {
-    assert_eq!(MAKETYPE(CV_8U, 2), CV_8UC2);
-    assert_eq!(MAKETYPE(CV_32S, 3), CV_32SC3);
-    assert_eq!(MAKETYPE(CV_64F, 1), CV_64FC1);
+    assert_eq!(CV_MAKETYPE(CV_8U, 2), CV_8UC2);
+    assert_eq!(CV_MAKETYPE(CV_32S, 3), CV_32SC3);
+    assert_eq!(CV_MAKETYPE(CV_64F, 1), CV_64FC1);
 }
 
 #[test]